[ADD] LocalStorage * Refactor the player
