+++ app/locales/en.json
@@ -20,6 +20,7 @@
   "length": "Length",
   "no playing music": "No playing music :(",
   "paste the url here": "Paste the URL here",
+  "play error": "An error ocured while playing the track",
   "playlists": "Playlists",
   "remove from playlist": "Remove from playlist",
   "reset up next": "Reset Up Next",
