+++ app/views/app_layout.js
@@ -1,6 +1,8 @@
 import Mn from 'backbone.marionette';
 import ContentView from './content';
-import PlayerView from './player';
+import PlayerControlsView from './playerControls';
+import PlayerControlsExtendedView from './playerControlsExtended';
+import PlayerTimelineView from './playerTimeline';
 import ToolbarView from './toolbar';
 import application from '../application'
 
@@ -11,18 +13,37 @@ const AppLayout = Mn.LayoutView.extend({
 
     el: '[role="application"]',
 
+    ui: {
+        player: 'audio'
+    },
+
     regions: {
         toolbar: '[role="toolbar"]',
         content: '[role="contentinfo"]',
-        player: '#player-container'
+        playerControls: '.play-controls',
+        playerControlsExtended: '.controls-extended',
+        playerTimeline: '.timeline'
     },
 
     onRender() {
+        // Init the audio html element
+        application.audio = this.ui.player.get(0);
+
         this.showChildView('content', new ContentView({
             model: application.appState
         }));
-        this.showChildView('player', new PlayerView());
         this.showChildView('toolbar', new ToolbarView());
+
+        // Player Views
+        this.showChildView('playerControls', new PlayerControlsView({
+            model: application.appState
+        }));
+        this.showChildView('playerControlsExtended', new PlayerControlsExtendedView({
+            model: application.appState
+        }));
+        this.showChildView('playerTimeline', new PlayerTimelineView({
+            model: application.appState
+        }));
     }
 });
