+++ app/views/templates/app_layout.jst
@@ -1,3 +1,8 @@
 <div role="toolbar"></div>
 <div role="contentinfo"></div>
-<div id="player-container"></div>
+<div class="player">
+    <div class="play-controls"></div>
+    <div class="timeline"></div>
+    <div class="controls-extended"></div>
+</div>
+<audio autoplay preload autobuffer></audio>
