--- app/views/templates/player.jst
@@ -1,60 +1,0 @@
-<audio id="player" autoplay preload autobuffer></audio>
-<div class="player">
-    <div class="play-controls">
-        <% if (false) { %>
-            <img class="album" src="album"></img>
-        <% } else { %>
-            <button class="album">
-                <svg>
-                    <use xlink:href="<%= require('../../assets/icons/music-lg.svg') %>"/>
-                </svg>
-            </button>
-        <% }%>
-
-        <button id="prev">
-            <svg>
-                <use xlink:href="<%= require('../../assets/icons/prev-lg.svg') %>"/>
-            </svg>
-        </button>
-        <button id="play">
-            <svg>
-                <use xlink:href="<%= require('../../assets/icons/play-lg.svg') %>"/>
-            </svg>
-        </button>
-        <button id="next">
-            <svg>
-                <use xlink:href="<%= require('../../assets/icons/next-lg.svg') %>"/>
-            </svg>
-        </button>
-    </div>
-    <div class="timeline">
-        <div class="text-container">
-            <p id="trackname"><%= t('no playing music') %></p>
-        </div>
-        <div id="progress-bar">
-            <div id="progress"></div>
-        </div>
-        <p id="current-time">0.00</p>
-        <p id="total-time">0:00</p>
-    </div>
-    <div class="controls">
-        <button id="repeat">
-            <svg>
-                <use xlink:href="<%= require('../../assets/icons/repeat-sm.svg') %>"/>
-            </svg>
-        </button>
-        <button id="speaker">
-            <svg>
-                <use xlink:href="<%= require('../../assets/icons/speaker-sm.svg') %>"/>
-            </svg>
-        </button>
-        <div id="volume-bar">
-            <div id="volume"></div>
-        </div>
-        <button id="broadcast">
-            <svg>
-                <use xlink:href="<%= require('../../assets/icons/broadcast-sm.svg') %>"/>
-            </svg>
-        </button>
-    </div>
-</div>
