+++ app/views/templates/playerTimeline.jst
@@ -1,0 +1,12 @@
+<div class="text-container">
+    <% if (!currentTrack) { %>
+            <p id="trackname"><%= t('no playing music') %></p>
+    <% } else { %>
+            <p id="trackname"><%= currentTrack.get('metas').title %></p>
+    <% }%>
+</div>
+<div id="progress-bar">
+    <div id="progress" style="width:<%= timePercent %>"></div>
+</div>
+<p id="current-time"><%= currentTime %></p>
+<p id="total-time"><%= totalTime %></p>
